+++ test/test.core_utils.sh
@@ -71,6 +71,17 @@ function teardown() {
 }
 
 
+@test "\`install' installs from gitlab" {
+  MSU_EXTERNAL_LIB="${BATS_TMPDIR}/gh"
+  source lib/core_utils.sh
+  samplemodule="GL:GochoMugo/msu-test"
+  run install "${samplemodule}"
+  [ "${status}" -eq 0 ]
+  echo "${output}" | grep "${sym_tick}"
+  [ -d "${MSU_EXTERNAL_LIB}/msu" ]
+}
+
+
 @test "\`install_from_list' installs from a list in a file" {
   MSU_EXTERNAL_LIB="${BATS_TMPDIR}/install-many"
   source lib/core_utils.sh